+++ packages/api-contract/src/types.ts
@@ -5,7 +5,7 @@ import type { ApiBase } from '@polkadot/api/base';
 import type { ApiTypes } from '@polkadot/api/types';
 import type { Text, u64 } from '@polkadot/types';
 import type { ContractExecResultResult, ContractSelector, StorageDeposit } from '@polkadot/types/interfaces';
-import type { Codec, TypeDef } from '@polkadot/types/types';
+import type { Codec, ICompact, INumber, TypeDef } from '@polkadot/types/types';
 import type { BN } from '@polkadot/util';
 import type { Abi } from '.';
 
@@ -73,11 +73,23 @@ export interface DecodedMessage {
 }
 
 export interface ContractOptions {
-  gasLimit?: bigint | string | number | BN;
+  gasLimit?: bigint | string | number | BN | WeightV2;
   storageDepositLimit?: bigint | string | number | BN | null;
   value?: bigint | BN | string | number;
 }
 
 export interface BlueprintOptions extends ContractOptions {
   salt?: Uint8Array | string | null;
+}
+
+export interface WeightV2 {
+  refTime: ICompact<INumber>;
+  proofSize: ICompact<INumber>;
+}
+
+export interface WeightAll {
+  v1Weight: BN;
+  v2Weight: {
+    refTime: BN;
+  };
 }